code: make a separate header for SALSSACodeGen
