+++ lib/Transforms/IPO/SALSSACodeGen.cpp
@@ -1,3 +1,4 @@
+#include "llvm/Transforms/IPO/SALSSACodeGen.h"
 #include "llvm/IR/BasicBlock.h"
 #include "llvm/IR/Dominators.h"
 #include "llvm/IR/InstIterator.h"
